requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-ollama
langgraph
faiss-cpu
sentence-transformers
python-dotenv
pypdf
